Interface created, option added in setting
