+++ main.ts
@@ -12,14 +12,12 @@ import {
 	Setting,
 	TFile,
 } from "obsidian";
-import {
-	GlobalSettings,
-	TaskBoardSettingTab,
-} from "./src/settings/TaskBoardSettingTab";
 
 import { AddTaskModal } from "src/components/AddTaskModal";
-import ConfigModal from "src/components/BoardModal";
+import ConfigModal from "src/settings/BoardModal";
+import { GlobalSettings } from "src/interfaces/KanbanView";
 import { KanbanView } from "./src/views/KanbanView";
+import { TaskBoardSettingTab } from "./src/settings/TaskBoardSettingTab";
 import { VIEW_TYPE_KANBAN } from "./src/views/KanbanView";
 // Import required modules
 import fs from "fs";
